refactor: merge FieldAnalyzer and TranslationFieldAnalyzer services
- Integrate all translation analysis functionality into FieldAnalyzer
- Convert TranslationFieldAnalyzer to a thin compatibility wrapper
- Optimize API to use single cache key and parallel operations
- Improve type safety with RawField interface
- Eliminate redundant method calls and database queries
- No breaking changes - full backward compatibility maintained
